Don't hardcode the bottom margin
The tooltip appears way to much above/below the related element. This is due to the hardcoded offset of 10px. Rather than hardcoding it, get it programmatically via $.css(). This PR adds two more variables to get hands on the arrow and its height and replaces the hardcoded offset by the calculated one. This also allows designers to override the offset (border-width) through css.